Remove unnecessary casts and toString() calls
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1808512 13f79535-47bb-0310-9956-ffa450edef68

+++ src/scratchpad/src/org/apache/poi/hwmf/usermodel/HwmfPicture.java
@@ -89,7 +89,7 @@ public class HwmfPicture {
             }
             
             consumedSize += wr.init(leis, recordSize, recordFunction);
-            int remainingSize = (int)(recordSize - consumedSize);
+            int remainingSize = recordSize - consumedSize;
             if (remainingSize < 0) {
                 throw new RecordFormatException("read too many bytes. record size: "+recordSize + "; comsumed size: "+consumedSize);
             } else if(remainingSize > 0) {
